Add user options to use an updated country file
Add command line options, -c, --cty_dat for specifying the location of
the updated country file.

Add check for XDX_CTY environment variable for specifying the location
of the updated country file.

When neither of the user set methods are used, check the preferences
directory for the existence of a country file and use it.

Fallback to the system installed version if any of the above are not
used or the specified file does not exist or is not a regular file.

Enlarge the buffer used to read in country file records and test the
size of the record read to avoid a buffer overrun.  Newer country files
have several large records including exception callsigns that would
cause a buffer overrun.

(cherry picked from commit cc1c1950cf72790875e124d1926fe9bd9fbfc057)