Use accessor methods rather than setting struct members directly
The GTK+3 migration guide specifies that compilation should be done
with:

	make CFLAGS+="-DGSEAL_ENABLE"

This environment variable identifies direct access of structure members.
These have been corrected.

In the case of the font selection dialog, the Apply button now appears
but it isn't processed by the code and acts similarly to the Cancel
button.  As the font selection dialog will be replaced in GTK+3, this
behavior will be left as is through the transition.