Use GCC #pragma to suppress warnings of deprecated functions
Employ GCC diagnostic pragma to suppress warnings on these "deprecated"
functions.  This absurdity is brought about due to the intent that CSS
is used for theming but if using already existing interfaces such as
pango font names the developer must come up with a solution to convert
such names to CSS.  The library does not have such support.

Source:
https://discourse.gnome.org/t/converting-fontchooser-font-style-to-css-font-style/6599

The Gtk3 devhelp states that gtk_widget_override_font() is deprecated
and that CSS should be useed instead.  There is no further information
on how this might be achieved and for those of us trying to convert from
pango font names to CSS and without knowledge of the internals of how
widgets parse such font names, doing so seems fraught with error and is
a waste of developer time.

Perhaps there is something out there that handles this conversion and is
well debugged but I have yet to find it.