Adjust the layout of the main GUI slightly
Created a new vertical box to hold the menu bar, function keys bar, main
text entry, and status bar separate from the paned window.  This made
setting various margins easier and limited the size of the text
highlighting sidebar.  This also allowed packing the windows a bit
differently to resolve the oversized width of the text highlighting
sidebar.

Added small margins to the main windows and function and side bars.